Clean up the routes
