--- app/routes/pages.season5.jsx
@@ -1,9 +1,0 @@
-export default function Season5() {
-    return (
-        <div>
-            <h2>Season 5</h2>
-            <div className="klaviyo-form-WrgVSp"></div>
-            <h1>Parth is gay</h1>
-        </div>
-    )
-}
